Move all the non-csv related bits out of the CSV module and into the file processor.
